% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_data.R
\name{bin_data}
\alias{bin_data}
\title{Generate histogram of the input vector}
\usage{
bin_data(x, lower_bound, upper_bound, n_bins = 1000, excess_bins = TRUE)
}
\arguments{
\item{x}{The privacy parameter rho}

\item{lower_bound}{The lower bound of the data}

\item{upper_bound}{The upper bound of the data}

\item{n_bins}{The number of bins. (Default: 1000)}

\item{excess_bins}{Do you want to include excess bins from -Inf to lower_bound and upper_bound to Inf? (Default: TRUE)}
}
\value{
Histogram of the data with corresponding x values.
}
\description{
Generate histogram of the input vector
}
