% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_cdf.R
\name{boot_cdf}
\alias{boot_cdf}
\title{Generate B bootstrapped cdfs from an input cdf}
\usage{
boot_cdf(cdf, B = 1000, project = FALSE)
}
\arguments{
\item{cdf}{A cdf. Ideally the output of projected_dp_cdf. If an unprojected noisy cdf is passed the projection will be done automatically.}

\item{B}{Number of bootstrap repetitions}
}
\value{
A isotone dp cdf of the data.
}
\description{
Generate B bootstrapped cdfs from an input cdf
}
