% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_bootstrap.R
\name{dp_bootstrap}
\alias{dp_bootstrap}
\title{Generic pipeline for dp bootstrap}
\usage{
dp_bootstrap(
  x,
  mechanism,
  statistics_of_interest,
  B = 1000,
  one_bootstrap_sample,
  bootstrap_sampler,
  bootstrap_loop
)
}
\arguments{
\item{x}{input data}

\item{mechanism}{dp mechanism. A function that takes x as input and produces "synthetic data".}

\item{statistics_of_interest}{Functions to calculate on the output of the mechanism, collected in a list}

\item{B}{number of bootstrap repetitions}

\item{bootstrap_sampler}{A function of how to sample bootstrap samples from the output of the mechanism.}
}
\value{
A list with the calculated statistics of interest and their bootstrap distribution
}
\description{
Generic pipeline for dp bootstrap
}
