% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_bootstrap_cdf.R
\name{dp_bootstrap_cdf}
\alias{dp_bootstrap_cdf}
\title{Wrapper function to generate dp cdf and bootstrapped cdfs from univariate input data x}
\usage{
dp_bootstrap_cdf(
  x,
  lower_bound,
  upper_bound,
  n_bins = 1000,
  excess_bins = TRUE,
  n = NULL,
  epsilon,
  delta,
  rho = NULL,
  B = 1000,
  project = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric input data}

\item{lower_bound}{The lower bound of the data}

\item{upper_bound}{The upper bound of the data}

\item{n}{Number of observations in original data (if public).}

\item{epsilon}{Privacy parameter epsilon}

\item{delta}{Privacy parameter delta}

\item{rho}{Privacy parameter rho (Either epsilon and delta or rho must be provided.)}

\item{B}{Number of bootstrap repetitions.}

\item{...}{Placeholder for potential additional arguments for the functions called within the wrapper.}
}
\value{
A dp cdf and bootstrap samples
}
\description{
Wrapper function to generate dp cdf and bootstrapped cdfs from univariate input data x
}
