% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_cdf.R
\name{dp_cdf}
\alias{dp_cdf}
\title{Generate differentially private cdf from histogram}
\usage{
dp_cdf(binned_data, n = NULL, epsilon, delta, rho = NULL)
}
\arguments{
\item{binned_data}{The output of bin_data}

\item{n}{The number of observations in the original data (if public, otherwise the max of the noisy cumsum will be used). Default: NULL (i.e. use maximum of noisy cumsum).}

\item{epsilon}{Privacy parameter epsilon}

\item{delta}{Privacy parameter delta}

\item{rho}{Privacy parameter rho. Either epsilon and delta or rho must be provided. Rho overwrites any entries in epsilon and delta.}
}
\value{
A noisy cdf of the data.
}
\description{
Generate differentially private cdf from histogram
}
