% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fast_factorize_M}
\alias{fast_factorize_M}
\title{Create factorization of sum matrix M}
\usage{
fast_factorize_M(n)
}
\arguments{
\item{n}{The length of the vector to privatize}
}
\value{
Matrix L (and/or R) where L\%*\%R = M
}
\description{
Provides a function that makes it easy to sample synthetic data from a Generator
}
