% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_inv_M}
\alias{get_inv_M}
\title{Create inverse of query matrix M for cumulative (prefix) sum queries}
\usage{
get_inv_M(n)
}
\arguments{
\item{n}{The length of the vector to privatize}
}
\value{
Inverse of query matrix M
}
\description{
Create inverse of query matrix M for cumulative (prefix) sum queries
}
