% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_M}
\alias{get_M}
\title{Create query matrix M for cumulative (prefix) sum queries}
\usage{
get_M(n)
}
\arguments{
\item{n}{The length of the vector to privatize}
}
\value{
Query matrix M
}
\description{
Create query matrix M for cumulative (prefix) sum queries
}
