% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_zcdp_rho}
\alias{get_zcdp_rho}
\title{Calculate the rho of zCDP based on epsilon and delta}
\usage{
get_zcdp_rho(epsilon, delta)
}
\arguments{
\item{epsilon}{The epsilon}

\item{delta}{The delta}
}
\value{
rho
}
\description{
Calculate the rho of zCDP based on epsilon and delta
}
