% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_zcdp_sigma}
\alias{get_zcdp_sigma}
\title{Calculate the sigma (standard deviation) for the Gaussian mechanism to achieve rho-zCDP}
\usage{
get_zcdp_sigma(rho, sensitivity = 1)
}
\arguments{
\item{rho}{The privacy parameter rho}

\item{sensitivity}{The sensitivity of the query}
}
\value{
rho
}
\description{
Calculate the sigma (standard deviation) for the Gaussian mechanism to achieve rho-zCDP
}
