% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_converter.R
\name{population_converter}
\alias{population_converter}
\title{Function to run Monte Carlo experiments}
\usage{
population_converter(population)
}
\arguments{
\item{population}{a dataset with a fixed population or a function to draw n values from a population}
}
\value{
A function to sample n observations from a population
}
\description{
Function to run Monte Carlo experiments
}
