% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dp_cdf.R
\name{project_dp_cdf}
\alias{project_dp_cdf}
\title{Project noisy cdf to isotone cdf}
\usage{
project_dp_cdf(dp_cdf)
}
\arguments{
\item{dp_cdf}{The output of dp_cdf}
}
\value{
A isotone dp cdf of the data.
}
\description{
Project noisy cdf to isotone cdf
}
