% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Function to run Monte Carlo experiments}
\usage{
run_experiment(
  run_id = "testrun",
  results_directory = NULL,
  population,
  n,
  mechanism,
  ...,
  get_synthetic_data,
  statistics_of_interest,
  B = 1000,
  store_P_tilde = FALSE,
  store_P_tilde_tilde = FALSE
)
}
\arguments{
\item{population}{a dataset with a fixed population or a function to draw n values from a population}

\item{n}{The sample size}

\item{mechanism}{dp mechanism. A function that takes a data set x as input and produces a generative model.}

\item{statistics_of_interest}{Functions to calculate the statistics of interest on the output of the mechanism, collected in a list}

\item{B}{number of bootstrap repetitions}

\item{store_P_tilde}{Do you want to store the initial generative model? Default: FALSE}

\item{store_P_tilde_tilde}{Do you want to sstore all bootstrapped generative models? Default: FALSE}
}
\value{
A list with the calculated statistics of interest and their bootstrap distribution
}
\description{
Function to run Monte Carlo experiments
}
